# Replace manual buffer creation with System.Buffers.ArrayPool