--- a/CliWrap/EventStream/CommandEvent.cs
+++ b/CliWrap/EventStream/CommandEvent.cs
@@ -1,5 +1,4 @@
-﻿using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
+﻿using System.Diagnostics.CodeAnalysis;
 
 namespace CliWrap.EventStream
 {