--- a/CliWrap/Internal/Extensions/StreamExtensions.cs
+++ b/CliWrap/Internal/Extensions/StreamExtensions.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System.Buffers;
+using System.Collections.Generic;
 using System.IO;
 using System.Runtime.CompilerServices;
 using System.Text;
@@ -12,46 +13,59 @@ namespace CliWrap.Internal.Extensions
         public static async Task CopyToAsync(this Stream source, Stream destination, bool autoFlush,
             CancellationToken cancellationToken = default)
         {
-            var buffer = new byte[BufferSizes.Stream];
-            int bytesRead;
-
-            while ((bytesRead = await source.ReadAsync(buffer, cancellationToken)) != 0)
+            var buffer = ArrayPool<byte>.Shared.Rent(BufferSizes.Stream);
+            try
             {
-                await destination.WriteAsync(buffer, 0, bytesRead, cancellationToken);
+                int bytesRead;
 
-                if (autoFlush)
-                    await destination.FlushAsync(cancellationToken);
+                while ((bytesRead = await source.ReadAsync(buffer, cancellationToken)) != 0)
+                {
+                    await destination.WriteAsync(buffer, 0, bytesRead, cancellationToken);
+
+                    if (autoFlush)
+                        await destination.FlushAsync(cancellationToken);
+                }
+            }
+            finally
+            {
+                ArrayPool<byte>.Shared.Return(buffer);
             }
         }
 
         public static async IAsyncEnumerable<string> ReadAllLinesAsync(this StreamReader reader,
             [EnumeratorCancellation] CancellationToken cancellationToken = default)
         {
-            var stringBuilder = new StringBuilder();
-
-            var buffer = new char[BufferSizes.StreamReader];
-            int charsRead;
-
-            while ((charsRead = await reader.ReadAsync(buffer, cancellationToken)) > 0)
+            var buffer = ArrayPool<char>.Shared.Rent(BufferSizes.StreamReader);
+            try
             {
-                for (var i = 0; i < charsRead; i++)
+                int charsRead;
+
+                var stringBuilder = new StringBuilder();
+                while ((charsRead = await reader.ReadAsync(buffer, cancellationToken)) > 0)
                 {
-                    if (buffer[i] == '\n')
+                    for (var i = 0; i < charsRead; i++)
                     {
-                        // Trigger on buffered input (even if it's empty)
-                        yield return stringBuilder.ToString();
-                        stringBuilder.Clear();
-                    }
-                    else if (buffer[i] != '\r')
-                    {
-                        stringBuilder.Append(buffer[i]);
+                        if (buffer[i] == '\n')
+                        {
+                            // Trigger on buffered input (even if it's empty)
+                            yield return stringBuilder.ToString();
+                            stringBuilder.Clear();
+                        }
+                        else if (buffer[i] != '\r')
+                        {
+                            stringBuilder.Append(buffer[i]);
+                        }
                     }
                 }
-            }
 
-            // Yield what's remaining
-            if (stringBuilder.Length > 0)
-                yield return stringBuilder.ToString();
+                // Yield what's remaining
+                if (stringBuilder.Length > 0)
+                    yield return stringBuilder.ToString();
+            }
+            finally
+            {
+                ArrayPool<char>.Shared.Return(buffer);
+            }
         }
     }
 }