--- a/CliWrap/Internal/Extensions/StreamExtensions.cs
+++ b/CliWrap/Internal/Extensions/StreamExtensions.cs
@@ -1,5 +1,4 @@
-﻿using System.Buffers;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.IO;
 using System.Runtime.CompilerServices;
 using System.Text;
@@ -13,59 +12,45 @@ namespace CliWrap.Internal.Extensions
         public static async Task CopyToAsync(this Stream source, Stream destination, bool autoFlush,
             CancellationToken cancellationToken = default)
         {
-            var buffer = ArrayPool<byte>.Shared.Rent(BufferSizes.Stream);
-            try
-            {
-                int bytesRead;
+            using var buffer = new PooledSharedBuffer<byte>(BufferSizes.Stream);
+            int bytesRead;
 
-                while ((bytesRead = await source.ReadAsync(buffer, cancellationToken)) != 0)
-                {
-                    await destination.WriteAsync(buffer, 0, bytesRead, cancellationToken);
-
-                    if (autoFlush)
-                        await destination.FlushAsync(cancellationToken);
-                }
-            }
-            finally
+            while ((bytesRead = await source.ReadAsync(buffer.Array, cancellationToken)) != 0)
             {
-                ArrayPool<byte>.Shared.Return(buffer);
+                await destination.WriteAsync(buffer.Array, 0, bytesRead, cancellationToken);
+
+                if (autoFlush)
+                    await destination.FlushAsync(cancellationToken);
             }
         }
 
         public static async IAsyncEnumerable<string> ReadAllLinesAsync(this StreamReader reader,
             [EnumeratorCancellation] CancellationToken cancellationToken = default)
         {
-            var buffer = ArrayPool<char>.Shared.Rent(BufferSizes.StreamReader);
-            try
-            {
-                int charsRead;
+            using var buffer = new PooledSharedBuffer<char>(BufferSizes.StreamReader);
+            int charsRead;
 
-                var stringBuilder = new StringBuilder();
-                while ((charsRead = await reader.ReadAsync(buffer, cancellationToken)) > 0)
+            var stringBuilder = new StringBuilder();
+            while ((charsRead = await reader.ReadAsync(buffer.Array, cancellationToken)) > 0)
+            {
+                for (var i = 0; i < charsRead; i++)
                 {
-                    for (var i = 0; i < charsRead; i++)
+                    if (buffer.Array[i] == '\n')
                     {
-                        if (buffer[i] == '\n')
-                        {
-                            // Trigger on buffered input (even if it's empty)
-                            yield return stringBuilder.ToString();
-                            stringBuilder.Clear();
-                        }
-                        else if (buffer[i] != '\r')
-                        {
-                            stringBuilder.Append(buffer[i]);
-                        }
+                        // Trigger on buffered input (even if it's empty)
+                        yield return stringBuilder.ToString();
+                        stringBuilder.Clear();
+                    }
+                    else if (buffer.Array[i] != '\r')
+                    {
+                        stringBuilder.Append(buffer.Array[i]);
                     }
                 }
+            }
 
-                // Yield what's remaining
-                if (stringBuilder.Length > 0)
-                    yield return stringBuilder.ToString();
-            }
-            finally
-            {
-                ArrayPool<char>.Shared.Return(buffer);
-            }
+            // Yield what's remaining
+            if (stringBuilder.Length > 0)
+                yield return stringBuilder.ToString();
         }
     }
 }