Related to #83

Running only `BasicBenchmarks`, with 50 randomly generated lines (current benchmark is at 10000 which seems a bit excessive)

Before:
```
BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.100-rc.1.20452.10
  [Host]     : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT
  DefaultJob : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT


|           Method |     Mean |    Error |   StdDev | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|---------:|---------:|------:|--------:|------:|------:|------:|----------:|
| RunProcessAsTask | 64.92 ms | 1.629 ms | 4.803 ms |  0.79 |    0.06 |     - |     - |     - | 100.58 KB |
|          Sheller | 79.87 ms | 2.502 ms | 7.138 ms |  0.95 |    0.07 |     - |     - |     - | 166.58 KB |
|   MedallionShell | 82.20 ms | 1.629 ms | 2.854 ms |  0.95 |    0.04 |     - |     - |     - | 164.99 KB |
|          CliWrap | 86.41 ms | 1.713 ms | 3.001 ms |  1.00 |    0.00 |     - |     - |     - | 383.38 KB |
```

After:
```
BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.100-rc.1.20452.10
  [Host]     : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT
  DefaultJob : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT


|           Method |     Mean |    Error |   StdDev | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|---------:|---------:|------:|--------:|------:|------:|------:|----------:|
| RunProcessAsTask | 59.76 ms | 1.191 ms | 1.746 ms |  0.69 |    0.05 |     - |     - |     - | 100.58 KB |
|          Sheller | 72.51 ms | 1.415 ms | 1.182 ms |  0.84 |    0.06 |     - |     - |     - | 166.57 KB |
|          CliWrap | 85.41 ms | 1.707 ms | 4.527 ms |  1.00 |    0.00 |     - |     - |     - | 143.31 KB |
|   MedallionShell | 88.07 ms | 2.151 ms | 6.239 ms |  1.03 |    0.10 |     - |     - |     - | 164.99 KB |
```