@Tyrrrz 2020-09-30
Looks great! Do you think it's possible to wrap these into some `IDisposable` type, so one can do something like this:

```csharp
using var pooledBuffer = ArrayPool.Rent<byte>(...);

// ...

while ((bytesRead = await source.ReadAsync(pooledBuffer.Array, cancellationToken)) != 0) { }
```

I'm thinking it would provide extra safety. Maybe something like that already exists?